chore: Moves CRUD components to the Datasource component
